+++ CHANGELOG.md
@@ -9,5 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update existing packages [#1](https://github.com/rokwire/rokwire-building-block-sdk-go/issues/1)
 
+### Added
+- Add common code from BB template [#3](https://github.com/rokwire/rokwire-building-block-sdk-go/issues/3)
+
 [Unreleased]: https://github.com/rokwire/rokwire-building-block-sdk-go/compare/v1.0.0....HEAD
 [1.0.0]: https://github.com/rokwire/rokwire-building-block-sdk-go/tree/v1.0.0